fix(Header comp): doLogOut func added
